refactor: move provider endpoints to modules
Moved these endpoints at once, as moving one pulled the others:
/v1/provider-attributes-schema
/v1/provider-regions
/v1/providers/{address}
/v1/auditors

refs #1272
