+++ apps/api/src/app.ts
@@ -23,6 +23,7 @@ import { leasesRouter } from "./deployment/routes/leases/leases.router";
 import { featuresRouter } from "./features/routes/features/features.router";
 import { healthzRouter } from "./healthz/routes/healthz.router";
 import { clientInfoMiddleware } from "./middlewares/clientInfoMiddleware";
+import { providerRegionsRouter } from "./provider/routes/provider-regions/provider-regions.router";
 import { providersRouter } from "./provider/routes/providers/providers.router";
 import { apiRouter } from "./routers/apiRouter";
 import { dashboardRouter } from "./routers/dashboardRouter";
@@ -43,6 +44,7 @@ import {
   stripePricesRouter,
   stripeWebhook
 } from "./billing";
+import { auditorsRouter, providerAttributesSchemaRouter } from "./provider";
 import { Scheduler } from "./scheduler";
 import { createAnonymousUserRouter, getAnonymousUserRouter } from "./user";
 
@@ -99,6 +101,9 @@ appHono.route("/", certificateRouter);
 appHono.route("/", featuresRouter);
 appHono.route("/", getBalancesRouter);
 appHono.route("/", providersRouter);
+appHono.route("/", auditorsRouter);
+appHono.route("/", providerAttributesSchemaRouter);
+appHono.route("/", providerRegionsRouter);
 
 appHono.route("/", healthzRouter);
