+++ apps/api/src/caching/helpers.ts
@@ -77,15 +77,12 @@ export const cacheKeys = {
   getProviderGraphData: "getProviderGraphData",
   web3IndexRevenue: "web3IndexRevenue",
   getProviderActiveLeasesGraphData: "getProviderActiveLeasesGraphData",
-  getProviderAttributesSchema: "getProviderAttributesSchema",
   getTemplates: "getTemplates",
   getMarketData: "getMarketData",
-  getAuditors: "getAuditors",
   getProviderList: "getProviderList",
   getTrialProviderList: "getTrialProviderList",
   getTrialRegisteredProviderList: "getTrialRegisteredProviderList",
   getChainStats: "getChainStats",
-  getProviderRegions: "getProviderRegions",
   getMainnetNodes: "getMainnetNodes",
   getTestnetNodes: "getTestnetNodes",
   getSandboxNodes: "getSandboxNodes",
