+++ apps/api/src/core/providers/http-sdk.provider.ts
@@ -4,6 +4,7 @@ import {
   BidHttpService,
   BlockHttpService,
   DeploymentHttpService,
+  GitHubHttpService,
   LeaseHttpService,
   ProviderHttpService
 } from "@akashnetwork/http-sdk";
@@ -14,3 +15,5 @@ import { apiNodeUrl } from "@src/utils/constants";
 const SERVICES = [BalanceHttpService, AuthzHttpService, BlockHttpService, BidHttpService, DeploymentHttpService, LeaseHttpService, ProviderHttpService];
 
 SERVICES.forEach(Service => container.register(Service, { useValue: new Service({ baseURL: apiNodeUrl }) }));
+
+container.register(GitHubHttpService, { useValue: new GitHubHttpService({ baseURL: "https://raw.githubusercontent.com" }) });
