+++ apps/api/src/provider/controllers/auditor/auditor.controller.ts
@@ -1,0 +1,12 @@
+import { singleton } from "tsyringe";
+
+import { AuditorService } from "@src/provider/services/auditors/auditors.service";
+
+@singleton()
+export class AuditorController {
+  constructor(private readonly auditorService: AuditorService) {}
+
+  async getAuditors() {
+    return await this.auditorService.getAuditors();
+  }
+}
