+++ apps/api/src/provider/controllers/provider-attributes-schema/provider-attributes-schema.controller.ts
@@ -1,0 +1,12 @@
+import { singleton } from "tsyringe";
+
+import { ProviderAttributesSchemaService } from "@src/provider/services/provider-attributes-schema/provider-attributes-schema.service";
+
+@singleton()
+export class ProviderAttributesSchemaController {
+  constructor(private readonly providetAttributesSchemaService: ProviderAttributesSchemaService) {}
+
+  async getProviderAttributesSchema() {
+    return await this.providetAttributesSchemaService.getProviderAttributesSchema();
+  }
+}
