+++ apps/api/src/provider/http-schemas/auditor.schema.ts
@@ -1,0 +1,11 @@
+import { z } from "zod";
+export const AuditorListResponseSchema = z.array(
+  z.object({
+    id: z.string(),
+    name: z.string(),
+    address: z.string(),
+    website: z.string()
+  })
+);
+
+export type AuditorListResponse = z.infer<typeof AuditorListResponseSchema>;
