+++ apps/api/src/provider/routes/index.ts
@@ -1,1 +1,4 @@
 export * from "@src/provider/routes/providers/providers.router";
+export * from "@src/provider/routes/provider-attributes-schema/provider-attributes-schema.router";
+export * from "@src/provider/routes/provider-regions/provider-regions.router";
+export * from "@src/provider/routes/auditors/auditors.router";
