+++ apps/api/src/provider/services/provider-attributes-schema/provider-attributes-schema.service.ts
@@ -1,0 +1,15 @@
+import { GitHubHttpService } from "@akashnetwork/http-sdk";
+import { minutesToSeconds } from "date-fns";
+import { singleton } from "tsyringe";
+
+import { Memoize } from "@src/caching/helpers";
+
+@singleton()
+export class ProviderAttributesSchemaService {
+  constructor(private readonly gitHubHttpService: GitHubHttpService) {}
+
+  @Memoize({ ttlInSeconds: minutesToSeconds(5) })
+  async getProviderAttributesSchema() {
+    return await this.gitHubHttpService.getProviderAttributesSchema();
+  }
+}
