+++ apps/api/src/types/provider.ts
@@ -101,46 +101,6 @@ export interface ProviderDetail extends ProviderList {
   }[];
 }
 
-export type ProviderAttributesSchema = {
-  host: ProviderAttributeSchemaDetail;
-  email: ProviderAttributeSchemaDetail;
-  organization: ProviderAttributeSchemaDetail;
-  website: ProviderAttributeSchemaDetail;
-  tier: ProviderAttributeSchemaDetail;
-  "status-page": ProviderAttributeSchemaDetail;
-  "location-region": ProviderAttributeSchemaDetail;
-  country: ProviderAttributeSchemaDetail;
-  city: ProviderAttributeSchemaDetail;
-  timezone: ProviderAttributeSchemaDetail;
-  "location-type": ProviderAttributeSchemaDetail;
-  "hosting-provider": ProviderAttributeSchemaDetail;
-  "hardware-cpu": ProviderAttributeSchemaDetail;
-  "hardware-cpu-arch": ProviderAttributeSchemaDetail;
-  "hardware-gpu": ProviderAttributeSchemaDetail;
-  "hardware-gpu-model": ProviderAttributeSchemaDetail;
-  "hardware-disk": ProviderAttributeSchemaDetail;
-  "hardware-memory": ProviderAttributeSchemaDetail;
-  "network-provider": ProviderAttributeSchemaDetail;
-  "network-speed-up": ProviderAttributeSchemaDetail;
-  "network-speed-down": ProviderAttributeSchemaDetail;
-  "feat-persistent-storage": ProviderAttributeSchemaDetail;
-  "feat-persistent-storage-type": ProviderAttributeSchemaDetail;
-  "workload-support-chia": ProviderAttributeSchemaDetail;
-  "workload-support-chia-capabilities": ProviderAttributeSchemaDetail;
-  "feat-endpoint-ip": ProviderAttributeSchemaDetail;
-  "feat-endpoint-custom-domain": ProviderAttributeSchemaDetail;
-};
-
-export type ProviderAttributeSchemaDetail = {
-  key: string;
-  type: "string" | "number" | "boolean" | "option" | "multiple-option";
-  required: boolean;
-  description: string;
-  values?: Array<ProviderAttributeSchemaDetailValue>;
-};
-
-export type ProviderAttributeSchemaDetailValue = { key: string; description: string; value?: string };
-
 export type Auditor = {
   id: string;
   name: string;
