+++ packages/http-sdk/src/index.ts
@@ -16,3 +16,4 @@ export * from "./deployment/deployment-http.service";
 export * from "./lease/lease-http.service";
 export * from "./provider/provider-http.service";
 export * from "./utils/isHttpError";
+export * from "./git-hub/git-hub-http.service";
